move sph kernels to own file; improve docstrings
