+++ doc/sections/subsections/dev-feec.rst
@@ -1,0 +1,39 @@
+FEEC
+----
+
+Derham complex
+^^^^^^^^^^^^^^
+
+.. automodule:: struphy.feec.psydac_derham
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
+
+Grids
+^^^^^
+
+.. automodule:: struphy.topology.grids
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
+
+Mass Operators (bilinear forms)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: struphy.feec.mass
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
+
+
+Projection of spline functions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: struphy.feec.basis_projection_ops
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
