+++ doc/sections/subsections/dev-pic.rst
@@ -1,0 +1,34 @@
+PIC
+---
+
+Particle base class
+^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: struphy.pic.base.Particles
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
+
+
+Particel-to-grid accumulation
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: struphy.pic.accumulation.particles_to_grid
+    :members:
+    :special-members:
+    :show-inheritance:
+
+
+Accumulation kernels
+^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: struphy.pic.accumulation.accum_kernels
+    :members:
+    :special-members:
+    :show-inheritance:
+
+.. automodule:: struphy.pic.accumulation.accum_kernels_gc
+    :members:
+    :special-members:
+    :show-inheritance:
