+++ doc/sections/subsections/dev-sph.rst
@@ -1,0 +1,50 @@
+SPH
+---
+
+
+.. _pusher_kernels_sph:
+
+Pusher kernels
+^^^^^^^^^^^^^^
+
+.. automodule:: struphy.pic.pushing.pusher_kernels_sph
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __in
+
+
+.. _coeff_kernels_sph:
+
+Coefficient kernels
+^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: struphy.pic.pushing.eval_kernels_sph
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
+
+
+.. _eval_kernels_sph:
+
+Evaluation kernels
+^^^^^^^^^^^^^^^^^^
+
+.. automodule:: struphy.pic.sph_eval_kernels
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
+
+
+.. _smoothing_kernels:
+
+Smoothing kernels
+^^^^^^^^^^^^^^^^^^
+
+.. automodule:: struphy.pic.sph_smoothing_kernels
+    :members:
+    :special-members:
+    :show-inheritance:
+    :exclude-members: __init__
