+++ src/struphy/propagators/push_vin_sph_pressure.py
@@ -9,7 +9,7 @@ from line_profiler import profile
 
 from struphy.io.options import LiteralOptions, OptionsBase
 from struphy.models.variables import SPHVariable
-from struphy.pic.pushing import eval_kernels_sph, pusher_kernels
+from struphy.pic.pushing import eval_kernels_sph, pusher_kernels_sph
 from struphy.pic.pushing.pusher import Pusher
 from struphy.propagators.base import Propagator
 from struphy.utils.pyccel import Pyccelkernel
@@ -23,15 +23,18 @@ class PushVinSPHpressure(Propagator):
 
     .. math::
 
-        \frac{\textnormal d \mathbf v_p(t)}{\textnormal d t} = \kappa_p \sum_{i=1}^N w_i \left( \frac{1}{\rho^{N,h}(\boldsymbol \eta_p)} + \frac{1}{\rho^{N,h}(\boldsymbol \eta_i)} \right) DF^{-\top}\nabla W_h(\boldsymbol \eta_p - \boldsymbol \eta_i) \,,
+        \frac{\textnormal d \mathbf v_p(t)}{\textnormal d t} = \mathbf g - \sum_{i=1}^N w_p w_i \left( \frac{\partial \mathcal U}{\partial \rho}(\boldsymbol \eta_p) + \frac{\partial \mathcal U}{\partial \rho}(\boldsymbol \eta_i) \right) DF^{-\top}\nabla W_h(\boldsymbol \eta_p - \boldsymbol \eta_i) \,,
 
-    where :math:`DF^{-\top}` denotes the inverse transpose Jacobian, and with the smoothed density
+    where :math:`\mathbf g` is a constant acceleration and the second term corresponds to the pressure gradient.
+    Here, :math:`\mathcal U(\rho)` denotes the internal energy per unit mass
+    as a function of the mass density :math:`\rho` and :math:`DF^{-\top}` denotes the inverse transpose Jacobian
+    arising in the pull back of the gradient of the smoothing kernel :math:`W_h` 
+    chosen from :mod:`~struphy.pic.sph_smoothing_kernels`. 
+    Two choices of the internal energy are implemented:
+    
+    * Isothermal closure: :math:`\mathcal U(\rho) = \kappa \, \ln(\rho)`, where :math:`\kappa` is constant.
+    * Polytropic closure: :math:`\mathcal U(\rho) = \kappa \, \rho^{\gamma - 1} / (\gamma - 1)`, where :math:`\kappa` is the polytropic constant and :math:`\gamma = C_p / C_v` is the polytropic index.
 
-    .. math::
-
-        \rho^{N,h}(\boldsymbol \eta) = \frac 1N \sum_{j=1}^N w_j \, W_h(\boldsymbol \eta - \boldsymbol \eta_j)\,,
-
-    where :math:`W_h(\boldsymbol \eta)` is a smoothing kernel from :mod:`~struphy.pic.sph_smoothing_kernels`.
     Time stepping:
 
     * Explicit from :class:`~struphy.ode.utils.ButcherTableau`
@@ -80,6 +83,9 @@ class PushVinSPHpressure(Propagator):
 
         gravity : tuple, default=(0.0, 0.0, 0.0)
             Constant gravity vector added in the SPH pressure push.
+        
+        kappa : float, default=1.0
+            Coefficient in the internal energy function.
 
         thermodynamics : {"isothermal", "polytropic"}, default="isothermal"
             Thermodynamic closure selecting the SPH pressure kernel.
@@ -93,6 +99,7 @@ class PushVinSPHpressure(Propagator):
         kernel_width: tuple = None
         algo: OptsAlgo = "forward_euler"
         gravity: tuple = (0.0, 0.0, 0.0)
+        kappa: float = 1.0
         thermodynamics: OptsThermo = "isothermal"
 
         def __post_init__(self):
@@ -153,9 +160,9 @@ class PushVinSPHpressure(Propagator):
 
         # pusher kernel
         if self.options.thermodynamics == "isothermal":
-            kernel = Pyccelkernel(pusher_kernels.push_v_sph_pressure)
+            kernel = Pyccelkernel(pusher_kernels_sph.push_v_sph_pressure)
         elif self.options.thermodynamics == "polytropic":
-            kernel = Pyccelkernel(pusher_kernels.push_v_sph_pressure_ideal_gas)
+            kernel = Pyccelkernel(pusher_kernels_sph.push_v_sph_pressure_ideal_gas)
 
         gravity = xp.array(self.options.gravity, dtype=float)
 
@@ -167,6 +174,7 @@ class PushVinSPHpressure(Propagator):
             kernel_nr,
             *self.options.kernel_width,
             gravity,
+            self.options.kappa,
         )
 
         # the Pusher class wraps around all kernels
